+++ results/UCF_AFTER_RESULTS.json
@@ -1,0 +1,8 @@
+{
+  "true_values": [0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0, 0, 1, 1, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0],
+  "scores": [0.15189523, 0.25026545, 0.36038864, 0.22268008, 0.8565128, 0.18306376, 0.15188594, 0.3298905, 0.51985943, 0.1670035, 0.34760967, 0.51975054, 0.14232656, 0.34746304, 0.7055121, 0.48649737, 0.46786496, 0.3272651, 0.4095528, 0.30418882, 0.25253287, 0.16101691, 0.6381502, 0.1289915, 0.16425905, 0.8775885, 0.69536287, 0.2889564, 0.25905627, 0.28935578, 0.2969002, 0.8691662, 0.5449785, 0.551839, 0.24923784, 0.78630143, 0.24160631, 0.2791857, 0.42169142, 0.91286457, 0.8485528, 0.13456152, 0.13077034, 0.34198308, 0.16443205, 0.11487684, 0.7456797, 0.6943358, 0.33060277, 0.5771214, 0.73160386, 0.6414973, 0.42065382, 0.5300108, 0.6370006, 0.22699, 0.40615577, 0.6253927, 0.42742717, 0.78792864, 0.17057657, 0.22522056, 0.26754326, 0.90559363, 0.5579381, 0.22991085, 0.6887664, 0.57749236, 0.17507955, 0.16135584, 0.84652966, 0.4615819, 0.23605183, 0.28863004, 0.8670522, 0.26185924, 0.70978457, 0.14418747, 0.7542475, 0.3426486, 0.87436926, 0.20853327, 0.82393855, 0.48720092, 0.18158075, 0.16088864, 0.19124207, 0.3147573, 0.88817835, 0.42370096, 0.22186215, 0.22955222],
+  "predictions": [0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 0.0, 1.0, 1.0, 1.0, 0.0, 1.0, 1.0, 0.0, 0.0, 1.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 1.0, 0.0, 1.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0],
+  "fpr": [0.0, 0.01694915, 0.01694915, 0.03389831, 0.03389831, 0.05084746, 0.05084746, 0.06779661, 0.06779661, 0.08474576, 0.08474576, 0.11864407, 0.11864407, 0.13559322, 0.13559322, 0.18644068, 0.18644068, 0.22033898, 0.22033898, 0.27118644, 0.27118644, 0.37288136, 0.37288136, 1.0],
+  "tpr": [0.0, 0.0, 0.09090909, 0.09090909, 0.24242424, 0.24242424, 0.42424242, 0.42424242, 0.45454545, 0.45454545, 0.51515152, 0.51515152, 0.60606061, 0.60606061, 0.84848485, 0.84848485, 0.90909091, 0.9090909, 0.93939394, 0.93939394, 0.96969697, 0.96969697, 1.0, 1.0],
+  "auc": 0.895737
+}
